implemented trigger pause line handler
